requirements.txt
PyQt6>=6.4.0
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
cryptography>=38.0.0
